Little script to create random samplings of GDML files.
How do you encode the uncertainty present in the geometry description of
an experiment? A systematic study of all geometric variables by scanning
over them in a linear fashion becomes quite cumbersome. So, here we take
a probabilistic many-worlds approach: we encode the uncertainty in the
gdml file and run a pre-processor to generate a random sampling of many
gdml files that is consistent with a 67% confidence level, given the
uncertainties encoded in the files.

But really, I just wanted a way to quickly create a number of geometries
that vary in certain parameters (uniformly).

How to use this? Use the following syntax anywhere in your files:
  random_normal(mean,sigma)
with values for mean and sigma. They will be replaced by a normally
distributed value when sample_gdml.sh runs over it. Also works with
  random_uniform(mean,half)
where half is half the range (i.e. -1.0 -- +1.0 has half = 1.0).

You can specify a file or directory. The preprocessed file or directory
will be placed in a new directory with randomized name.

You can also pipe a stream into this command and it will return a stream
for you to redirect somewhere.

Works not just on gdml files, of course.