Save the actual variables in sample_gdml.vars in the temp dir.
The assumption here is that the variables will not change position
for the set of geometries that you are creating. Filesystem and bash
will ensure that for a single run, but of course this can't be ensured
for different runs of sample_gdml that have different variables enabled.

Since sample_gdml does not assign names or other unique IDs to the vars
this is an inherent limitation that would require significant work to
fix. Providing the values this way will hopefully be sufficient.

To use for simulations, by the way:
  scripts/sample_gdml.sh geometry_sculpt
  cd sample_gdml.2KrmpE/
  ../build/remoll ../macros/runexample.mac
This will use the master remoll executable and macro, but use the
geometry that is in the copied geometry_sculpt directory with the
custom geometry parameters.