Implemented auxiliary tag DetType to put optical photon tracking inside the GDML file instead of the generic detector class.
Possible values right now are "opticalphoton", "lowenergyneutral",
and "charged" (default). These are not mutually exclusive options,
as for example the PMTs are both "opticalphoton" and "lowenergyneutral"
since optical photons are low energy neutrals.

Surely there's a better way to solve this properly, but it works for
now with sufficient flexibility for remoll.