New event generator based on external file
This event generator takes a remollout.root file as input.  Why is that
useful? Simulations of the electrons getting through the Pb wall are very
slow (since now many electrons make it through). So, we have a virtual
detector that is located right after the Pb wall. We can now, once, run
a long simulation of electrons, and use that as input for optical simulations
of only the downstream pion detector system.