Split up pion lucite detector rotation
pionDetectorLuciteSectorAngleBase is +90*deg for PMT towards outside,
and -90*deg for PMT towards inside. The remaining angle due to non-z
orientation of the prevailing pion 'winds' can now be put in a single
variable pionDetectorLuciteSectorAnglePitch ~ 5*deg. This ensures that
the same lucite plan is always first, regardless of inside or outside
orientation.