New function remollDetCon::GetPhysicalVolumes
Get the list of physical volumes associated to a certain logical volume.
Because the tree structure in geant4 geometry is from mother logical to
daughter physical to daughter logical, there is no way to get from
mother logical to mother physical without a full tree transversal.

The efficiency of this function is suboptimal (many lists get created
and deleted as it transcends the tree), but who cares since this is
going to be called only once during initialization anyway.