Moved from CLHEP::Random to G4Random
This change is a requirement for multithreading to work (but don't get
your hopes up). This fixes issue #20.

The #defines like G4RandFlat are available as of geant4 version 9.4.p02,
so don't try to compile with version of geant4 older than 10 years
(though there are plenty of other reasons to upgrade).