Move commands from remollMessenger into parent classes
G4GenericMessenger allows for much more streamlined definition of
commands that just change class members or call simple methods. This
makes ownership of the commands clearer. That's in particular useful for
classes that are multiplied in multithreaded operation (primary
generator action and most other actions, with the exception of run
action). Instead of passing pointers around to remollMessenger, the
necessary messengers are now generated as needed by the action classes
(and other classes).