BeamTarget now owned by VEventGen as well
Unfortunately BeamTarget is a mix of actual data and calculational
framework (MS) so it is not clear how to treat it exactly. May need to
be surgically pried open. For now, placed it in the event generator. The
global target volume treatment is turned into statics (sympton of the
problem, if it requires this as a solution)