Delete old event generator before creating new one (fixes #28)
In the primary event generator we start off with a default moller event
generator. When/if a new generator is defined, this default event
generator is replaced. Previously the previous event generator was not
deleted. The event generator messenger associated with the (now
inactive) previous event generator maintained precedence over the new
active event generator. Therefore none of the commands took effect on
the new event generator.

Fixed bug report #28.