Set the names of event generators
In general these are the same as the names with which they are set in
/remoll/gen.

A bit of additional care had to be taken with the aluminum generator
since one class serves three event generators (elasticAl, inealsticAl,
quasielasticAl). Some creative use of static maps variables and static
map-creating functions lead to a moderately elegant solution. If you
call the constructor with a type outside of the valid range, you'll
either get an out_of_range exception or an empty name. If you want to
know which one it is, you'll have to try it because I didn't.