Explicitly use moller evgen for test_raster macro
The test_raster macro really only tests whether the raster works
properly. test_general.C therefore now creates a few more plots.
However, if no event generator is explicitly set in the macro,
it is not clear what event generator is actually used. In the MT
branch this default is set to the moller generator. Here the macro
is updated to explicitly set the test_raster generator to the moller
event generator as well. This differs from earlier default behavior
in master and v1.0.0 (which is somewhat ill-defined), but agrees
with future default behavior (no event generator specified -> moller).