Non-emtpy argument inherited class needs to initialize base class.
Since remollTextFile inherits from TObject, it needs to initialize
the TObject members properly when not using a constructor with the
same signature. In this case remollTextFile(const char* fn) needs
a TObject() initialization.

In similar vein, the assignment operator needs to call the TObject
class operator= before doing its thing.

Finally, virtual destructor ensures that the TObject destructor is
called when remollTextFile is deleted.