Change how physcis list verbose levels are handled.
Now the default is to just shut up about physics lists when starting
up. It is up to the user to decide if they want that output. In most
cases it may not be useful to users, but having it available in the
test log files may allow us to pick up differences from version to
version.