Modification to create a macro command to disable SDs.
Each sensitive detector gets a /remoll/SD/det<detno>/enable command
that can be used to disable and enable that sensitive detector at
run-time. This could be used to reduce the number of hits written
to the output files. By default all SDs are enabled. Possibly a
default state can be added in GDML through an auxiliary tag, e.g.
<auxiliary auxname="SensDetEnable" auxvalue="false">, again with
a default of true.

New branch to experiment with this.