Define new InitializeMaterials to do proper kryptonite detection
Previously the constructor of UserSteppingAction defined the list of
kryptonite materials by looping over the materials database. That's
before any materials are defined, however, so the list remained empty.

With a static local flag in the main UserSteppingAction, we now init
the list when the first track is used. That's guaranteed to have the
geometry at that point.

Caveat. Multiple runs with changes in geometry will have the same
kryptonite materials list. I.e. if you do change the material of
something from a kryptonite to a non-kryptonite material between
two /run/beamOn commands, you'll get wrong results in the second call.