Add 4 tracking planes at for now incorrect positions
Should be at:
    22.635 m
    23.635 m
    27.635 m
    28.635 m
per David. But that last one is past the detector 28 model of the
main detector array. That doesn't make sense. Either 28 needs to
move or we have to move the tracking planes.

So, for now I put the tracking planes a bit close together:
<define>
  <position name="trackingPlaneFront1_position" z="-2.5" unit="m"/>
  <position name="trackingPlaneFront2_position" z="-2." unit="m"/>
  <position name="trackingPlaneBack1_position"  z="+2." unit="m"/>
  <position name="trackingPlaneBack2_position"  z="+2.5" unit="m"/>
</define>
instead of the required
<define>
  <position name="trackingPlaneFront1_position" z="-3" unit="m"/>
  <position name="trackingPlaneFront2_position" z="-2." unit="m"/>
  <position name="trackingPlaneBack1_position"  z="+2." unit="m"/>
  <position name="trackingPlaneBack2_position"  z="+3" unit="m"/>
</define>
to agree with David.