Turn /remoll/evgen/command into statics
Because multiple event generators are created and all use the same,
e.g., /remoll/evgen/thmin, the correct value does not get passed along
to the right generators (only to the moller generator because it is
created first). Turning the values static allows us to most easily keep
the same commands and change, e.g., thmin for all generators at the same
time.