Modified logic for determining whether a hit is by the first step in volume
This uses the pre step status, and if undefined the stored post step status.

Possible times when this could go wrong? None that I can imagine right now,
but maybe there are cases when the pre step status is fUndefined (typically
the very first step) AND there is no stored post step status. But that is
what we had before too. So, I think we're good now.