Partial rewrite of remollGenBeam, building on @cameronc137 and #180
Now raster is perpendicular to beam direction, can also be set as a
vector, but there are more generic flat and gaussian origin position
profiles than can be chosen with specified spread.

You can now do something like:
```
/remoll/evgen/beam/origin 0 0 0 cm
/remoll/evgen/beam/originspread 10 10 0 cm
/remoll/evgen/beam/xmodel flat
/remoll/evgen/beam/ymodel gauss
```

Particularly useful for individual detector tests.

Not implemented: spread on direction (because it is difficult to
imagine how to maintain normalization, so the spread would have to
be on the theta and phi of the direction, not on the dx/dy/dz unit
vectors).