Bugfix to issue #172. Now kryptonite works again.
Since this is a quite major set of changes to how some of the hits
are being stored, I am putting this in a bugfix branch first so I
can test optical photon hit behavior and let @cameronc137 take a look.

In any case, here's hit behavior:
- by default only hits are stored that satisfy the following:
  - charge particle, energy > 0.1 MeV
  - anywhere along a primary track
  - first hit in volume for secondary track
- all hits (even those not stored) are added to the detector sums

To modify this behavior you can use
```
/remoll/SD/type 4001 lowenergyneutral
/remoll/SD/type 4001 secondaries
```
along with similar GDML statements (also possible: `opticalphotons`
and `boundaryhits`).

As for kryptonite, this was all moved to remollDetectorConstruction
and stepping action was removed (since now not useful). Here is the
code for kryptonite:
```
/remoll/kryptonite/verbose 1
/remoll/kryptonite/add G4_Pb
/remoll/kryptonite/list
/remoll/kryptonite/enable
/remoll/kryptonite/disable
```
(This is different from previous `/remoll/kryptonite/set true` since
we need a method and those don't work with `G4GenericMessenger::DefineMethod`
since a G4bool looks like a G4int and yada yada yada.)

Finally, kryptonite now acts by setting all volumes that have the
specified material in question with userlimits to (0,0,0,inf,inf).
This means that all energy is deposited in the first hit, making
the sums correct. Essentially it sets the userlimits to:
```
/remoll/geometry/userlimits/usermaxallowedstep krypt_logical 0.0*mm
/remoll/geometry/userlimits/usermaxtracklength krypt_logical 0.0*mm
/remoll/geometry/userlimits/usermaxtime krypt_logical 0.0*ns
/remoll/geometry/userlimits/userminekine krypt_logical 1000*GeV
/remoll/geometry/userlimits/userminrange krypt_logical 1000*km
```
for those volumes. It does mean that steplimiter must be enabled
by default. And if you disable steplimiter and still try to use
userlimits or kryptonite there is no warning... so don't do that.