MG5 doesn't seem to understand fixed target mode, so use half target mass
For massless particles 1 and 2, the Mandelstam s is defined as 2 p1 p2.
- For colliders this turns into 4 E1 E2.
- For fixed targets this turns into 2 m E.

Unfortunately MG5 requires that one specifies E1 and E2 and this is always
used to calculate s = 4 E1 E2 using the collider formula.

So, this commit gets MG5 to use the right value of s by telling it that we
are colliding a massless beam of 11 GeV electrons with a massless beam of
255.5 keV electrons. No, it doesn't make sense. It does, however, produce
the correct cross sections and, presumably, pT distributions. Stay tuned.