Replace det type "all" with ability to specify >1 GDML DetType
This is a better solution than adding another label that essentially
does the same as what labels we already had. The reason why "all" was a
quick fix is because we overwrote every DetType with the last one when
reading the GDML file. I've reorganized how the GDML tags are read (yay
for c++ lambdas and automatics type deduction...)

@chandabindu I'm going to merge this back into develop. Can you first
let me know if it allows you to see all hits from both lowenergyneutrals
and secondaries?