If not validating GDML or checking overlaps, just hide the output
In the spirit of "all output is important" the numerous messages
from the GDML are more noise than useful. Only turn them on when
the overlap checks or schema validation is requested.

This essentially hijacks the G4cout stream by setting the failbit
to true until the stream is cleared again.