Bugfix to #233 where generated pion energy fails assertion
The problem was that the beam can radiate so much that the energy is
lower than the minimum energy with which we wish to generate. In that
case G4RandFlat::shoot(emin, emax) receives a larger emin than emax,
and interprets this as a backwards interval. This subsequently fails
the e > emin test.

In this case, when emax is the radiated energy, and emax < emin, we
now set emin = emax and generate in that interval. This also changes
the phase space allocated to this event to zero (so properly weighted
outputs should not see this event anymore).