Warn about incorrect field maps that could result in gaps in phi.
```
Warning /home/wdconinc/git/remoll/src/remollMagneticField.cc line 293
File map_directory/upstreamSymmetric_sensR_0.0.txt header contains a gap in the phi range which seems
to correspond perfectly with a slice in phi. This will result in a gap in coverage.
```
which is not printed for map_directory/upstreamSymmetric_sensR_0.1.txt