Change the default build type to RelWithDebInfo
The default build type was Debug, which is about ~30% slower than
RelWithDebInfo, and RelWithDebInfo contains the same backtrace info.
This will make code faster (since typically run without changing the
default build type).

A specific build type can be chosen without the neeed to specify a
command line flag by creating a file .buildtype (which is ignored by
git) with as content e.g. Debug. This will persist between deletes
of the build directory (but changing the file will not cause a new
build type to take effect, this is only read on the first call to
cmake in an empty directory). The content of the file can still be
overriden with a command line argument.