Use G4Rand instead of CLHEP::Rand
With the move to MT support, the treatment of random numbers was
modified in geant4. One should not call the CLHEP::Rand functions but
instead use the G4Rand versions which are MT safe. This mainly affects
the G4Random seeding, but consistency is probably preferred here.

See e.g.
https://twiki.cern.ch/twiki/bin/view/Geant4/QuickMigrationGuideForGeant4V10#Random_numbers.