Avoid overlap with 75% less poly at front end, remove padding
The main goal here is to get this merged in. The concrete shielding of
the target region is overlapping with the upstream spectrometer poly
shielding by about 75 cm. This commit fixes the overlap in an adhoc way.

Rather than removing concrete DS shielding on the target, I've chosen to
remove US shielding on the upstream spectrometer in the hope that this
will have less impact. A proper study is obviously needed.