Remove the explicit GeV^-2 units in A_PV(Al) generator
When we moved to a single G_F definition (in units GeV^-2) instead
of a magical number without units, I forgot to remove the GeV^-2
from the asymmetry formula.

Double checked that this is not affecting any other places where
G_F is used.

@cipriangal Please test to see if this fixes the issue discussed
in the comments of 0c3fe73beefd38d04239a09ac33f5d911427173c