Fix to the jlab.sh script that works in existing environments
When using singularity we don't call JLab CE's jlab.sh with an
empty environment (nor do we want to, we need $DISPLAY, for one).
This fixes the tcsh to bash translator for environment variables
to be more accommodating of environments with existing functions
(bash completion, module system, spack, you name it).