Add an implementation of a sieve with a few holes (and the unit 'in')
The demo sieve has three constant diameter holes to demonstrate
increasingly complex behavior:
- hole 1: central phi, parallal to z
- hole 2: offset in phi, parallel to z
- hole 3: offset in phi, at angle with z
All sectors have the same holes right now.

Some remaining issues: the angle with z is as if the hole was centered
in phi, even if the offset is non-zero. I was struggling with making
that work correctly and gave up for now.