/remoll/geometry/absolute_position <volume_name> (x,y,z)
Set of new macro commands to move stuff around
- /remoll/geometry/absolute_position <volume_name> (x,y,z)
- /remoll/geometry/relative_position <volume_name> (x,y,z)
- /remoll/geometry/absolute_rotation <volume_name> (x,y,z)
- /remoll/geometry/relative_rotation <volume_name> (x,y,z)

The absolute versions set the position/rotation inside the mother volume
(so NOT in world coordinates). The relative versions set the
position/rotation with respect to the current position/rotation.

Rotations have to be specified in x, y, z rotations (in that order). I
hope that's sufficient.

Geant4 does not support units on macro commands with multiple arguments,
so this has to be specified in mm and deg.