Raster is at -19.81 m now, which affects pos/angle correlations.
The reference position of the raster (i.e. where the raster is the
specified 5 mm x 5 mm size), is now again the center of the target,
which is going to be right for most solid targets too and is anyway
less important for the much smaller pos/angle correlations. For a
5 mm x 5 mm raster and 0.065 mrad/mm correlations, the raster is only
0.2 mm larger downstream than upstream for a 1.25 m long target.