Recently, having LIBGL_ALWAYS_INDIRECT set on the farm causes crashes
This notifies the users when LIBGL_ALWAYS_INDIRECT is set (to any value),
and provides info on how to unset the variable.

Historical note: In about 2016/2017 this environment variable was used
to fix a visualization crash on the farm. Many may still have lines in
their login scripts to set this variable.