Add part of flange upstream of collar 0
Note that the upstream end of this volume is limited by the extent of the upstream torus region and not the actual flange length. The upstream torus region has a half length of 3310 mm and centered at 2000 mm in global coordinates. So, in the global coordinates, this upper most point is at 2000-3310=-1310 mm.